+++ mm/kmalloc.h
@@ -2,6 +2,6 @@
 #define KMALLOC_H
 #include <stdint.h>
 void kmalloc_init();
-void *kmalloc(uint32_t size);
+void *kmalloc(uint64_t size);
 void kfree(void *ptr);
 #endif
